math_site/requirements.txt
numpy==1.23.5
django==4.1.4
drf-spectacular==0.24.2
djangorestframework==3.14.0